refactor(app): folder restructuring
