+++ frontend/src/main.jsx
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom/client'
 import { QueryClient, QueryClientProvider } from "react-query";
 import { Axios } from "axios";
 
-import App from "./pages/app/App";
+import App from "./App";
 import './index.css'
 
 const queryClient = new QueryClient();
